Game Lobby, Create Game, and View Game
Super basic functionality for game related functionality is working.
The game lobby shows games that you are running and playing in,
you can create a new game, and view games by id.
Game Lobby will show you the DM's photo for games you are playing in.

When saving or updating, the DAO will assign the ID to the Game Object.